+++ docs/source/public_api/utils.rst
@@ -7,3 +7,9 @@ Utility Functions.
     :no-inheritance-diagram:
     :no-heading:
     :no-main-docstr:
+
+probnum.utils.linalg
+--------------------
+
+.. automodapi:: probnum.utils.linalg
+    :no-heading:
