+++ src/probnum/utils/linalg/__init__.py
@@ -1,0 +1,3 @@
+from ._cholesky_updates import cholesky_update, triu_to_positive_tril
+
+__all__ = ["cholesky_update", "triu_to_positive_tril"]
